pyproject.toml
[project]
name = "qna-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.3.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
]
